Fix SQL related server error on dataset pages after second deployment on same provisioned infrastructure (#208)
PR to fix issue #830

## Problem

Immediately after provisioning the infrastructure, on a gitlab pipeline, when pressing a button to trigger a deployment job for staging (``sd_gigadb``) or live (``ld_gigadb``), the application is successfully deployed the first time. If we trigger the same job a second time, the application is still deployed but the resulting web site show a server error with the following SQL related error when navigating to a dataset page.
It doesn't matter whether we rebuild the image (by triggering a build job for that environment) before triggering  the second deployment job or not.

```
CDbCommand failed to execute the SQL statement: SQLSTATE[42601]: Syntax error: 7 ERROR: zero-length delimited identifier at or near """" LINE 1: ... "t1_c3" FROM "file_format" "format" WHERE ("format".""=$1)
^. The SQL statement executed was: SELECT "format"."id" AS "t1_c0", "format"."name" AS "t1_c1", "format"."description" AS "t1_c2", "format"."edam_ontology_id" AS "t1_c3" FROM "file_format" "format" WHERE ("format".""=:ypl0). Bound with :ypl0=37
```


## Cause 

It's because of those lines from ``gigadb-deploy-jobs.yml`` related to running database migrations:


```
    - docker-compose --tlsverify -H=$REMOTE_DOCKER_HOST -f ops/deployment/docker-compose.production-envs.yml run --rm  application ./protected/yiic custommigrations dropconstraints
    - docker-compose --tlsverify -H=$REMOTE_DOCKER_HOST -f ops/deployment/docker-compose.production-envs.yml run --rm  application ./protected/yiic custommigrations dropindexes
    - docker-compose --tlsverify -H=$REMOTE_DOCKER_HOST -f ops/deployment/docker-compose.production-envs.yml run --rm  application ./protected/yiic migrate --migrationPath=application.migrations.schema --interactive=0
```

The first time a deployment job is triggered, the migration level is not recorded, and the schema migration runs and complete after the database constraints and indexes have been dropped.
The second time a deployment job, the migration level is now recorded in the database so the schema migration is not run but the database constraints are still dropped,
so the database is in a state where foreign constraints are not defined which will cause join queries (like those performed by Yii lookup functions) to fail.

## Fix


One option is to reset migration level: since all the constraints need to be dropped, we need to reset migration level, so that the database schema migration can run and re-establish them.

However we are losing the benefit of using migrations, which is the incremental update of the schema as necessary and required by changes to the application.

The proper approach is to perform schema migrations and dropping of constraints and indexes if and only if there are new migrations to run.

In order to do that, we save into a variable (``MIGRATION_STATUS``) the output of executing ``yiic migrate new`` and then prefix the three lines above with a test (``[[ -z "`echo $MIGRATION_STATUS | grep "up-to-date"`" ]]``) on the output that will allows their execution only if the string saying that the migrations are up-to-date is not present in that variable.