Add setting for tests to pass captcha & scenario for submitting the form
Updated nginx config for dev and CI only to have their environment available in $_SERVER['GIGADB_ENV']

Updated UserController.php to test for that variable and if it matches dev or CI,
we rewrite the current session's captcha to be the value passed with the form.
That way we have both the  automated test can pass verification and the backend validation still happening

Created a feature specific (i.e non-generic) step for ensuring the test data is deleted from database
in CuratorSteps.php.

Added form filling generic steps in AcceptanceTest.php.

Wrote a second scenario in NewUser.feature to test that the form can be filled in and submitted successfully.