Add Codeception functional tests for captcha loading
Replace coverage-incomplete functional test for captcha
with new Codeception functional tests that cover more and relevant behaviours of
Captchas:

* the tests cover both captcha usage, the one on contact form and the one on user form.
* we test that the generated captcha is different between successive call
* no need to test for the image size specific to each form, as captcha image presentation
should be the same across usage. We are not harmonising in this task, it will be done later when we fix the big captcha issue
* we test the captcha is image/png mime which revealed a minor bug in User form controller (discrepancy between file format and generated format) which we fix here

This test is made more necessary now that acceptance tests scenario for user creation form is bypassing captcha using server variables
and environment detection, so we need to make sure captcha mechanism has test coverage.