Decouple debug mode from caching control
Make sure the existing YII_DEBUG, YII_TRACE_LEVEL and the new DISABLE_CACHE
can be configured like any other variables using gitlab variables or .env
and the ``generate_config.sh`` script.

Also ensure we have sensible default values depending on which environment we are in
(dev/CI or staging/live) for all the above variables. This is done in ``generate_config.sh``
(and also in ``gigadb-build-jobs.yml`` for the YII_DEBUG setting for staging).

The invalidation query is now defined in a configuration file (``main.php.dist``) for convenience.

Updated docs with an explainer about those variables.