Revert changes for bastion script to talk to Pg container on dockerhost
We no longer need to have  a process on bastion that connects to a legacy postgresql container on
the main ec2 vm for webapps,

So the fils-url-updater changes to accommodate running bare on the bastion vm (like downgrading PHP 7.4 -> 7.2 and therefore removing strong typing on instance variables) and connecting to postgres container
on webapp ec2 vm are no longer needed.

We don't need to open an additional port in the security group either.