Use exact version for behat/mink to avoid upgrade of symfony/browser-kit
because upgrade of that composer package triggers upgrade of those components:
symfony/browser-kit
symfony/css-selector
symfony/dom-crawler

Those three packages will introduce a new dependent package symfony/polyfill-php80

which causes problem with CActiveRecord.php  when loading test fixture for the tests/unit/SampleTest.php.

that test normally loads fixture for the attribute table which has ``protected/models/Attribute.php`` as the corresponding
ActiveRecord class.
But since the package upgrade, it is an equally non-namespaced  Attribute.php class in the package  symfony/polyfill-php80 which is loaded instead.

This is a stop-gap fix so we can commit the composer.lock.

(the proper fix is to replace all our Yii1 model with Yii2 model classes which are namespaced)