Fix logic for removing a custom image and its metadata for a dataset
A problem in the "beforeDelete" hook in Image prevented the actual record deletion to be performed
(missed the return true statement).

Also I've decided not to delete the physical directly. Instead we store the url to delete in a new table (images_todelete)
that will be housekept separately by a backend worker in the future.

Also makes sure that the JSON response from the remove image action is honest and contain a success is true statement
only if all the interactive steps of the removal action have been performed successfully (reasasigment to generic image,
deletion of the old image record and creation of a new record in the images_todelete table).

Updated the functional test to test the new logic.