Change approach for getting acceptance tests running on apple silicon
the previous approach didn't work on Intel based mac,

so instead of replacing the docker compose service for webdriver, I've add an additional one
specific to arm and kept the old one for intel based

then updated the up.sh script to select which one to instantiate based on current architecture.

Also reverted composer.lock and composer.json.dist to the version from develop for now. We will verify whether that need to change