fix: remove the include deprecation warning when executing playbooks
It's an issue in the role-secure-docker-daemon third party role.
I fixed it by forking the role (after creating a github issue there),
making a fix on my fork and update the link in the requirements.yml file to
point to my fork.
If it work for all of us, I'll make a PR to the Ansible repository of the original role